Stop video decoding when video element is hidden
With the background audio feature, media elements with audio and video
content will continue to play.  However the video portion will not be
displayed.  To reduce power used, the video decoding is paused.  The
feature is enabled by default and can be disabled by the command line
flag --disable-video-pausing.  This feature current only supports
progressive download content.  Video decoding for content using the
Media Source Extensions will not be paused.

Change-Id: Ifcea4bd29a885eb89d9641a2961fd8825f05ec5a